+++ web/src/js/components/contentviews/HttpMessage.tsx
@@ -11,6 +11,7 @@ import Button from "../common/Button";
 import CodeEditor from "./CodeEditor";
 import LineRenderer from "./LineRenderer";
 import ViewSelector from "./ViewSelector";
+import { copyViewContentDataToClipboard, fetchApi } from "../../utils";
 
 type HttpMessageProps = {
     flow: HTTPFlow;
@@ -31,9 +32,13 @@ export default function HttpMessage({ flow, message }: HttpMessageProps) {
         () => setMaxLines(Math.max(1024, maxLines * 2)),
         [maxLines],
     );
-    const [edit, setEdit] = useState<boolean>(false);
-    let url;
-    if (edit) {
+    const [isEdited, setIsEdited] = useState<boolean>(false);
+    const [isCopied, setIsCopied] = useState<boolean>(false);
+    const [isFetchingFullContent, setIsFetchingFullContent] =
+        useState<boolean>(false);
+
+    let url: string;
+    if (isEdited) {
         url = MessageUtils.getContentURL(flow, message);
     } else {
         url = MessageUtils.getContentURL(
@@ -45,7 +50,7 @@ export default function HttpMessage({ flow, message }: HttpMessageProps) {
     }
     const content = useContent(url, message.contentHash);
     const contentViewData = useMemo<ContentViewData | undefined>(() => {
-        if (content && !edit) {
+        if (content && !isEdited) {
             try {
                 return JSON.parse(content);
             } catch (e) {
@@ -60,11 +65,38 @@ export default function HttpMessage({ flow, message }: HttpMessageProps) {
         }
     }, [content]);
 
-    if (edit) {
+    const handleClickCopyButton = () => {
+        const url = MessageUtils.getContentURL(flow, message, contentView); //without the 'maxLines' parameter, so we can get the full content of the content view
+
+        setIsFetchingFullContent(true);
+
+        fetchApi(url)
+            .then((response) => {
+                if (!response.ok) {
+                    throw new Error(
+                        `${response.status} ${response.statusText}`.trim(),
+                    );
+                }
+                return response.json();
+            })
+            .then((data: ContentViewData) => {
+                copyViewContentDataToClipboard(data);
+                setIsCopied(true);
+                setTimeout(() => setIsCopied(false), 2000);
+            })
+            .catch((e) => {
+                console.error(e);
+            })
+            .finally(() => {
+                setIsFetchingFullContent(false);
+            });
+    };
+
+    if (isEdited) {
         const save = async () => {
             const content = editorRef.current?.getContent();
             await dispatch(flowActions.update(flow, { [part]: { content } }));
-            setEdit(false);
+            setIsEdited(false);
         };
         return (
             <div className="contentview" key="edit">
@@ -79,7 +111,7 @@ export default function HttpMessage({ flow, message }: HttpMessageProps) {
                     </Button>
                     &nbsp;
                     <Button
-                        onClick={() => setEdit(false)}
+                        onClick={() => setIsEdited(false)}
                         icon="fa-times text-danger"
                         className="btn-xs"
                     >
@@ -97,8 +129,19 @@ export default function HttpMessage({ flow, message }: HttpMessageProps) {
             <div className="contentview" key="view">
                 <div className="controls">
                     <h5>{desc}</h5>
+                    {contentViewData && contentViewData?.lines.length > 0 && (
+                        <Button
+                            onClick={handleClickCopyButton}
+                            icon="fa-clipboard"
+                            className="btn-xs"
+                            disabled={isFetchingFullContent}
+                        >
+                            {isCopied ? "Copied!" : "Copy"}
+                        </Button>
+                    )}
+                    &nbsp;
                     <Button
-                        onClick={() => setEdit(true)}
+                        onClick={() => setIsEdited(true)}
                         icon="fa-edit"
                         className="btn-xs"
                     >